+++ db/seeds/how_to_create_a_user.rb
@@ -6,6 +6,6 @@
 #         email: 'user@example.com',
 #         password_hash: BCrypt::Password.create('user_password')
 #       }
-# repo = UserRepo.new
+# repo = DevelopmentApp::UserRepo.new
 # repo.create(:users, usr)
 raise 'This is not for running!'