+++ lib/labels/views/label/properties.rb
@@ -23,6 +23,7 @@ module Labels
               form.add_field :language
               form.add_field :category
               form.add_field :sub_category
+              form.add_field :variable_set
             end
           end