+++ routes/labels/publish.rb
@@ -15,11 +15,19 @@ class LabelDesigner < Roda
         # (If coming from "BACK" button, show plain section using cached list of targets
       end
 
-      r.get 'show_targets' do
+      r.get 'callback_for_targets' do
         res = interactor.publishing_server_options
         # stash....
         if res.success
-          { content: render_partial { Labels::Publish::Batch::SelectTargets.call(res.instance) } }.to_json
+          show_in_callback { Labels::Publish::Batch::SelectTargets.call(res.instance) }
+        elsif res.instance[:timeout]
+          show_in_callback(content: 'Unable to get information - the server took too long to respond.',
+                           content_style: :error,
+                           error: res.message)
+        elsif res.instance[:refused]
+          show_in_callback(content: 'Unable to get information - the server might not be running.',
+                           content_style: :error,
+                           error: res.message)
         else
           show_json_error(res.message, status: 200)
         end
@@ -27,34 +35,28 @@ class LabelDesigner < Roda
 
       r.post 'select_labels' do
         res = interactor.select_targets(params[:batch]) #=> validate resolution
-        # if res.success
         show_page { Labels::Publish::Batch::SelectLabels.call(res.instance) }
-        # else
-        #   re_show_form(r, res, url: '/labels/publish/batch') do
-        #     Labels::Publish::Batch::Targets.call(form_values: params[:batch],
-        #                                          form_errors: res.errors)
-        #   end
-        # end
         # VALIDATE: 1) printer chosen; 2) Server chosen, 3) Chosen server is configured for the chosen printer.
         # Store params in step (selected targets)
         # Get list of ELIGIBLE labels (approved) with publish history (Show date updated - max published date as days since published)
       end
 
       r.post 'publish' do
+        # Validate: cannot publish multi without sub-labels...
         res = interactor.save_label_selections(multiselect_grid_choices(params))
         show_page { Labels::Publish::Batch::Publish.call(res.instance) }
       end
 
-      r.get 'send' do
+      r.get 'callback_for_send' do
         res = interactor.publish_labels # (store.read(:lbl_publish_steps))
         { content: render_partial { Labels::Publish::Batch::Send.call(res) } }.to_json
       end
 
       r.get 'feedback' do
-        res = interactor.publishing_status # (store.read(:lbl_publish_steps))
-        # TODO: differentiate between failure and exception - exception should stop polling...
+        res = interactor.publishing_status
         if res.success
-          payload = { content: res.instance.body.to_s, continuePolling: !res.instance.done }
+          content = render_partial { Labels::Publish::Batch::PublishState.call(res.instance) }
+          payload = { content: content, continuePolling: !res.instance.done }
           payload[:finaliseProgressStep] = 'cbl-current-step' if res.instance.done
           { updateMessage: payload }.to_json
         else