+++ CHANGELOG.md
@@ -10,6 +10,16 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Changed
 ### Fixed
 
+## [0.3.5] - 2019-01-24
+### Added
+- Dataminer module included.
+- Created by and Updated by on labels.
+- Labels can be archived and un-archived (active/inactive).
+- Publishing of labels is logged and uses the job queue to do the work.
+### Changed
+- Status logging for create, link and delete of labels.
+- List of variables for a label can be configured via the variable set and shared_config.
+
 ## [0.3.0] - 2019-01-04
 ### Added
 - Action on labels grid: `Refresh preview values from sub-labels` - combines preview values from a multi-label's sub-labels, replacing sample_data for the mult-label.